Avoid deprecated encode_point() method
- This allows `python-sshpubkeys` to work with `cryptography>=39.0.0`.
- `ec.EllipticCurvePublicNumbers.encode_point()` had been deprecated for
a long time, and was removed in commit pyca/cryptography@2b6e463 , or
`cryptography==39.0.0`.
- I manually tested to confirm that for all of the existing test cases
that called `to_string()`, this new `public_bytes()` method yielded the
same bytes as the old `encode_point()`.
- Since the "compressed" `to_string()` was already using
`public_bytes()` with `Encoding` and `PublicFormat`, it seems unlikely
that this change will break support for any versions of `cryptography`